Move packages out of src directory
