+++ cassandra/session.go
