+++ rsyslog_server.go
